+++ packages/core/cairo_alpha/src/generate/custom.ts
@@ -2,7 +2,8 @@ import type { CustomOptions } from '../custom';
 import type { AccessSubset } from '../set-access-control';
 import { resolveAccessControlOptions } from '../set-access-control';
 import { infoOptions } from '../set-info';
-import { upgradeableOptions } from '../set-upgradeable';
+import type { UpgradeableSubset } from '../set-upgradeable';
+import { resolveUpgradeableOptionsSubset } from '../set-upgradeable';
 import { generateAlternatives } from './alternatives';
 import type { MacrosSubset } from '../set-macros';
 import { resolveMacrosOptions } from '../set-macros';
@@ -11,6 +12,7 @@ const booleans = [true, false];
 
 type GeneratorOptions = {
   access: AccessSubset;
+  upgradeable: UpgradeableSubset;
   macros: MacrosSubset;
 };
 
@@ -19,7 +21,7 @@ function prepareBlueprint(opts: GeneratorOptions) {
     name: ['MyContract'],
     pausable: booleans,
     access: resolveAccessControlOptions(opts.access),
-    upgradeable: upgradeableOptions,
+    upgradeable: resolveUpgradeableOptionsSubset(opts.upgradeable),
     info: infoOptions,
     macros: resolveMacrosOptions(opts.macros),
   };